\docType{data}
\name{carbon_intensities}
\alias{carbon_intensities}
\title{Carbon intensity data}
\format{A data frame with 13 rows and 2 columns}
\source{
  Data come from
  \url{https://www.gov.uk/government/organisations/department-of-energy-climate-change/series/fuel-mix-disclosure-data-tables},
  \url{http://www.earth.org.uk/note-on-UK-grid-CO2-intensity-variations.html},
  and
  \url{https://www.gov.uk/government/publications/2012-greenhouse-gas-conversion-factors-for-company-reporting}.
}
\description{
  A data frame giving the carbon intensity of different
  electricity generating sources including electricity
  flows from interconnections with other European
  countries. The values are measured in g CO2/kWh.
}
\keyword{datasets}

