\docType{package}
\name{decctools}
\alias{decctools}
\alias{decctools-package}
\title{decctools}
\description{
  Provides easy access to United Kingdom energy statistics
  for R.  Most of the underlying data comes from the
  Department of Energy and Climate Change website,
  \url{http://www.decc.gov.uk}.
}
\seealso{
  \code{\link{get_LAD_data}}, \code{\link{get_MSOA_data}},
  \code{\link{get_LSOA_data}},
  \code{\link{get_lookup_table}}
}

