% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_area_classifications}
\alias{get_area_classifications}
\title{Gets Output Area classifications}
\source{
\url{http://www.ons.gov.uk/ons/guide-method/geography/products/area-classifications/ns-area-classifications/ns-2011-area-classifications/index.html}
}
\usage{
get_area_classifications()
}
\value{
a data frame giving the new LAD code and the most common
Supergroup classification name
}
\description{
Gets 2011 Output Area classification data from Office for National
Statistics.  At the moment, these are only available at ward level
and so the representative classification for the local authority is
simply taken as the most common ward classification.
}
\note{
In all of the other Office of National Statistics products
used in this package, Northern Ireland local authorities are not
assigned a "new" style code.  However in the area classifications,
they do have this new code but corresponding to differently named
authorities.  As a result, no area classifications are returned
for NI towns.
}

