% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_climate_data}
\alias{get_climate_data}
\title{Gets climate data for the local authorities}
\usage{
get_climate_data()
}
\value{
a data frame giving the new code and heating and cooling degree days.
}
\description{
Gets heating and cooling degree day data for each local authority.
This is based on the benchmarking paper.
}

