% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_code_lookup}
\alias{get_code_lookup}
\title{Gets a look-up between old and new LA codes}
\source{
Office for National Statistics
}
\usage{
get_code_lookup()
}
\value{
a data frame giving the old and new lookup codes, as well
as country and name
}
\description{
Gets a look-up table between old and new Local Authority codes.
For Northern Ireland, we cheat and reuse the old codes as new
codes to make later merging easy.
}

