% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_geo_lookup}
\alias{get_geo_lookup}
\title{Gets a lookup table to match IDs across geographies}
\usage{
get_geo_lookup()
}
\value{
a data frame
}
\description{
Provides a lookup table giving the local authority name and
corresponding IDs for LAD, MSOA, and LSOA geographies.
}
\note{
As SOA energy data are not available for Northern Ireland,
this method only returns a lookup for England, Wales, and
Scotland.  Also these lookup tables are based on the 2011
geographies.  When fetching data for other years, users may find
that codes don't align.  The recommended strategy is to merge on
the name where possible.
}

