\name{get_grid_mix}
\alias{get_grid_mix}
\title{Gets mix of fuel sources in UK electricity}
\source{
  The underlying data come from BM Reports,
  \url{http://www.bmreports.com/bsp/bsp_home.htm}, and are
  used under license from Elexon. However the BM Reports
  website is hard to use and so this function uses archived
  data compiled by the Renewable Energy Foundation at
  \url{http://www.ref.org.uk/fuel/}.
}
\usage{
  get_grid_mix(start, end)
}
\arguments{
  \item{start}{the start date for retrieving data.}

  \item{end}{the end date for retrieving data}
}
\value{
  a data frame with a datetime stamp and average power
  generated by fuel type in MW
}
\description{
  Gets the mix of fuel sources used to generate UK grid
  electricity between specified dates.  Valid dates are
  yyyy-mm-dd and must be between 2009-01-01 and the current
  day; the start must come before the end date.
}

