% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_LAD_metadata}
\alias{get_LAD_metadata}
\title{Gets metadata to describe LADs}
\usage{
get_LAD_metadata(dir)
}
\arguments{
\item{dir}{an optional directory in which to save a copy of the metadata}
}
\value{
a data frame giving metadata for each LAD.  See \link{LAD_metadata}
}
\description{
Gets metadata to decribe the local authority districts
}

