\name{get_LAD_population}
\alias{get_LAD_population}
\title{Gets the 2010 population estimates for all local authority districts}
\usage{
  get_LAD_population()
}
\value{
  a data frame with the LAU1_code and population
}
\description{
  This is a convenience function that returns the
  population data from \code{\link{LAD_metadata}}.
}
\seealso{
  \code{\link{LAD_metadata}}
}

