% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_last_date}
\alias{get_last_date}
\title{Gets the most recent date for available data}
\usage{
get_last_date()
}
\value{
a Date object.  If the REF website is unavailable, it will
return today's date on the assumption that once it's back up, it
will be up-to-date.
}
\description{
Gets the date of the last update to the REF fuels mix database.
}

