% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_LSOA_metadata}
\alias{get_LSOA_metadata}
\title{Gets metadata for all LSOAs}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/socio-economic-data-for-mlsoa-igz-and-llsoa-electricity-and-gas-estimates}
}
\usage{
get_LSOA_metadata(dir)
}
\arguments{
\item{dir}{an (optional) directory in which to save the downloaded
data}
}
\value{
a data frame with the LSOA id code, population, area (in
hectares), and number of households
}
\description{
Gets the socio-demographic data associated with each Lower Super
Output Area (LSOA). These data only cover England and Wales.
}

