\name{get_master_LSOA_params_list}
\alias{get_master_LSOA_params_list}
\title{Builds a master set of parameters for MSOA data}
\usage{
  get_master_LSOA_params_list(dir)
}
\arguments{
  \item{dir}{the directory in which to store a copy of the
  data}
}
\description{
  Creates a list of various parameters needed to download
  and extract MSOA data from the DECC website.
}

