% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_params_list}
\alias{get_params_list}
\title{Builds a master set of parameters for SOA data}
\usage{
get_params_list(level)
}
\arguments{
\item{level}{one of "LSOA" or "MSOA" specifying the output area
level}
}
\value{
a list containing the parameters necessary to read each
LSOA data file
}
\description{
Creates a list of various parameters needed to download and
extract SOA data from the DECC website.
}

