% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_population}
\alias{get_population}
\title{Downloads population estimates}
\source{
Office for National Statistics
\url{http://www.ons.gov.uk/ons/rel/census/2011-census/population-and-household-estimates-for-the-united-kingdom}
}
\usage{
get_population()
}
\value{
a data frame with the LAD id, name, population, and area (ha)
}
\description{
Downloads population estimates
}

