% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_SOA_metadata}
\alias{get_SOA_metadata}
\title{Gets metadata for Super Output Areas}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/socio-economic-data-for-mlsoa-igz-and-llsoa-electricity-and-gas-estimates}
}
\usage{
get_SOA_metadata(level, dir)
}
\arguments{
\item{level}{one of "LSOA" (default) or "MSOA" specifying the
output area level}

\item{dir}{an (optional) directory in which to save the downloaded
data}
}
\value{
a data frame with the SOA id code, population, area (in
hectares), and number of households
}
\description{
Gets the socio-demographic data associated with each Super Output
Area.
}

