% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_SOA_years}
\alias{get_SOA_years}
\title{Gets the years for which SOA data are available}
\usage{
get_SOA_years(level = "LSOA")
}
\arguments{
\item{level}{one of "LSOA" (default) or "MSOA" specifying the
output area level}
}
\value{
a numeric vector of valid years for \code{level}
}
\description{
Gets the years for which SOA data are available
}

