% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_urban_rural_classes}
\alias{get_urban_rural_classes}
\title{Gets urban/rural classifications}
\usage{
get_urban_rural_classes()
}
\value{
a data frame giving the urban rural classification for
each local authority with its old id code
}
\description{
Gets the urban/rural classification of each local authority.  The
data are from 2009.
}
\note{
The urban/rural classifications from the 2011 Census are not
yet available; see
\url{http://www.ons.gov.uk/ons/guide-method/geography/products/area-classifications/2011-rural-urban/index.html}.
The data used here are from
\url{http://www.ons.gov.uk/ons/guide-method/geography/products/area-classifications/rural-urban-definition-and-la/rural-urban-local-authority--la--classification--england-/index.html}
}

