\name{get_valid_years}
\alias{get_valid_years}
\title{Determines which year values are valid}
\usage{
  get_valid_years(wb, unit = "GWh")
}
\arguments{
  \item{wb}{the workbook to search}

  \item{unit}{the measurement unit of interest.  Defaults
  to GWh}
}
\value{
  a vector of years with data in the spreadsheet
}
\description{
  Determines which year values are valid
}

