% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{params}
\alias{params}
\title{SOA parameters}
\format{a data frame with 8 columns giving the parameters needed
to read the Excel spreadsheets containing the energy data}
\description{
Provides a table of parameters for downloading SOA energy
consumption data.  Used by \link{get_SOA_data}.
}
\keyword{datasets}

