% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_raw_SOA_data}
\alias{parse_raw_SOA_data}
\title{Parses raw SOA data}
\usage{
parse_raw_SOA_data(level, params)
}
\arguments{
\item{level}{one of "LSOA" or "MSOA" specifying the
output area level}

\item{params}{a list giving the function parameters including}
}
\value{
a data frame with the energy data.  NULL if not all of the
required parameters are specified
}
\description{
Parses the raw SOA data for a set of given parameters.  This
function deals with the fact that every DECC spreadsheet has a
slightly different layout, but the extraction is the same.
Namely, open up an Excel spreadsheet, get the right tab, extract
the data from the right columns, tidy, and return.
}

