% Generated by roxygen2 (4.0.1): do not edit by hand
\name{process_tab}
\alias{process_tab}
\title{Loads data from a specified tab}
\usage{
process_tab(wb, y, params)
}
\arguments{
\item{wb}{the open workbook containing all of the tabs}

\item{y}{the year of data to fetch}

\item{params}{a list of parameters giving the target id, sector,
and fuel}
}
\value{
a data frame with the data
}
\description{
Loads energy data from a specified tab in the LAD workbook.
}

