% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{scotland_igz}
\alias{scotland_igz}
\title{Scotland geography data}
\format{a data frame with 3 columns}
\description{
Provides a lookup between Scottish districts and their Scottish
Executive codes.  Used internally to build the
\link{get_geo_lookup} table.
}
\keyword{datasets}

