% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validate_directory}
\alias{validate_directory}
\title{Validates a user-specified directory}
\usage{
validate_directory(dir)
}
\arguments{
\item{dir}{the directory name}
}
\value{
the validated directory name
}
\description{
Ensures that a user-specified directory exists.  If the argument is
missing then a placeholder directory is created in $TEMP.
}

