% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validate_SOA_level}
\alias{validate_SOA_level}
\title{Validates a SOA level}
\usage{
validate_SOA_level(level)
}
\arguments{
\item{level}{a character vector of unprocessed SOA level values}
}
\value{
a single valid level value
}
\description{
Only one entry is allowed, either LSOA or MSOA
}

