% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_decc_data}
\alias{clean_decc_data}
\title{Cleans the raw DECC energy consumption data for easier use}
\usage{
clean_decc_data(df)
}
\arguments{
\item{df}{the raw DECC data frame from the spreadsheet}
}
\value{
a long database with headers by sector and fuel
}
\description{
Cleans the raw DECC energy consumption data for easier use
}

